Debug en cours

+++ Julia/capacityOfLeg.jl
@@ -4,7 +4,7 @@ using DataFrames
 
 function capacityOfLeg(PATH)
   file = readtable(PATH)
-  cap = Dict{UTF8String,Int64}()
+  cap = Dict{UTF8String,Float64}()
   for i = 1:size(file[1])[1]
     cap[join([file[1][i], file[3][i]], "-")] = file[6][i]
   end
